requirements.txt
# PyTorch (install via conda for CUDA support, or use pip with appropriate CUDA version)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Machine learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Progress bar
tqdm>=4.65.0

# Testing
pytest>=7.4.0
hypothesis>=6.82.0

# Documentation & PPT
python-pptx>=0.6.21

# Logging
tensorboard>=2.14.0

# Optional: for legacy TensorFlow code (not required for PyTorch version)
# tensorflow>=2.13.0
# tensorflow_addons>=0.21.0
